+++ gamestonk_terminal/stocks/quantitative_analysis/factors_model.py
