+++ gamestonk_terminal/stocks/technical_analysis/README.md
@@ -44,8 +44,6 @@ This menu aims to perform a technical analysis on a pre-loaded ticker chart, and
 [VOLATILITY](#VOLATILITY)
   * [bbands](#bbands)
     - bollinger bands
-  * [capm](#capm)
-    - capm model
 
 [VOLUME](#VOLUME)
   * [ad](#ad)
@@ -184,12 +182,6 @@ usage: aroon [-l N_LENGTH] [-s N_SCALAR] [-o N_OFFSET]
 usage: bbands [-h] [-l N_LENGTH] [-s N_STD] [-m S_MAMODE] [-o N_OFFSET]
 ```
 ![bbands](https://user-images.githubusercontent.com/25267873/108602984-28699580-739d-11eb-9b82-2683a9840145.png)
-
-## capm  <a name="capm"></a>
-```
-usage: capm
-```
-![capm]
 
 
 # VOLUME <a name="VOLUME"></a>
